backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.11.0
python-multipart==0.0.22
pdfplumber==0.11.9
python-docx==1.1.2
requests==2.32.3
google-genai==1.60.0
python-dotenv==1.0.1
pypdf==6.6.2
huggingface_hub==0.28.1